Issue #1302 Merge branch '5-Data_Delivery' into development
Conflicts:
    edexOsgi/build.edex/deploy-common/external-rules.xml
    edexOsgi/build.edex/deploy-common/plugin-methods.xml
    edexOsgi/com.raytheon.edex.plugin.text/src/com/raytheon/edex/plugin/text/dao/TextDao.java
    edexOsgi/com.raytheon.uf.edex.database/src/com/raytheon/uf/edex/database/plugin/PluginDao.java
    edexOsgi/com.raytheon.uf.edex.esb.camel/src/com/raytheon/uf/edex/esb/camel/Executor.java
    edexOsgi/build.edex/build.xml
    edexOsgi/build.edex/esb/conf/modes.xml
    edexOsgi/com.raytheon.edex.feature.uframe/feature.xml
    edexOsgi/com.raytheon.edex.plugin.grib/res/spring/grib-ingest.xml
    edexOsgi/com.raytheon.edex.plugin.grib/src/com/raytheon/edex/plugin/grib/Grib1Decoder.java
    edexOsgi/com.raytheon.edex.plugin.grib/src/com/raytheon/edex/plugin/grib/GribDecoder.java
    edexOsgi/com.raytheon.uf.common.gridcoverage/src/com/raytheon/uf/common/gridcoverage/GridCoverage.java


Former-commit-id: 7d0e4a6e75c157072d5c17918d88411b5a9e5e98 [formerly 8e69107b5afe12539b5fcc9ac6039b704951216f] [formerly 98d7beae5064edc8a52f374a4fffb6c98b5c2585] [formerly 31d7d09c10c841cfd0ac019699249840668963bc [formerly 98d7beae5064edc8a52f374a4fffb6c98b5c2585 [formerly 89583d97c8823818f87dc1a234078d85ba7bea4a]]]
Former-commit-id: 31d7d09c10c841cfd0ac019699249840668963bc
Former-commit-id: 465044dd2793505eba62cb46e6b3a8ac48623765 [formerly bcf575966b550519e01a35bd8091f358b5d36800]
Former-commit-id: e7a0f96a9a12671a5090f9ecae8a4d6332929ed7

+++ ncep/gov.noaa.nws.ncep.edex.plugin.stormtrack/unit-test/gov/noaa/nws/ncep/edex/plugin/stormtrack/util/StormTrackParserTest.java
@@ -1,8 +1,7 @@
 package gov.noaa.nws.ncep.edex.plugin.stormtrack.util;
 
-import static org.junit.Assert.*;
-import gov.noaa.nws.ncep.common.dataplugin.stormtrack.*;
-import gov.noaa.nws.ncep.edex.plugin.stormtrack.util.StormTrackParser;
+import static org.junit.Assert.assertEquals;
+import gov.noaa.nws.ncep.common.dataplugin.stormtrack.StormTrackRecord;
 
 import java.util.Calendar;
 
@@ -19,23 +18,23 @@ public class StormTrackParserTest {
 	public void testProcessLatLon() {
 		Float latLon;
 		latLon = StormTrackParser.processLatLon("975W");
-		assertEquals (-97.5, latLon);
+        assertEquals(-97.5, latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("1605E");
-		assertEquals (160.5, latLon);
+        assertEquals(160.5, latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("1605W");
-		assertEquals (-160.5, latLon);
+        assertEquals(-160.5, latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("623E");
-		assertEquals (62.3, latLon);
+        assertEquals(62.3, latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("62N");
-		assertEquals (6.2, latLon);
+        assertEquals(6.2, latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("847S");
-		assertEquals (-84.7, latLon);
+        assertEquals(-84.7, latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("847G");
-		assertEquals (999999., latLon);
+        assertEquals(999999., latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon("");
-		assertEquals (999999., latLon);
+        assertEquals(999999., latLon.doubleValue());
 		latLon = StormTrackParser.processLatLon(" ");
-		assertEquals (999999., latLon);
+        assertEquals(999999., latLon.doubleValue());
 	}
 
 	@Test
